pyproject.toml
[project]
name = "sonify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.51.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
]
